cdk8s_manifests/cdk8s_manifests/lib/helm.py
from typing import Dict, Any, Optional
from constructs import Construct
from cdk8s import Chart
import yaml
import os

class HelmChart(Construct):
    """Wrapper for Helm charts in cdk8s+."""
    
    def __init__(self, scope: Construct, id: str, 
                 chart_name: str, 
                 release_name: Optional[str] = None,
                 repository: Optional[str] = None,
                 version: Optional[str] = None,
                 namespace: Optional[str] = None,
                 values: Dict[str, Any] = None):
        """
        Initialize a Helm chart.
        
        Args:
            scope: The parent construct
            id: The construct ID
            chart_name: Name of the Helm chart
            release_name: Name for the Helm release (defaults to id)
            repository: Helm repository URL
            version: Chart version
            namespace: Kubernetes namespace
            values: Values to override in the Helm chart
        """
        super().__init__(scope, id)
        self.chart_name = chart_name
        self.release_name = release_name or id
        self.repository = repository
        self.version = version
        self.namespace = namespace
        self.values = values or {}
        
        # In a real implementation, this would call Helm via subprocess
        # and import the generated manifests.
        # For this example, we're just setting up the structure.
        
    def to_values_file(self, path: str) -> str:
        """Write values to a YAML file."""
        os.makedirs(os.path.dirname(path), exist_ok=True)
        with open(path, 'w') as f:
            yaml.dump(self.values, f)
        return path